rename supabase to data-fetcher
