+++ src/globals/data-fetcher.ts
